+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SphericalSurface3D: line_intersections, linesegment_intersections.
 - Sweep with muitiform profile contour.
 - OpenShell3D: from_faces (using faces graph)
+- SphericalFace3D: from_contours3d_and_rectangular_cut
 
 ### Fixed
 - ClosedShell3D: is_face_inside, get_subtraction_valid_faces, valid_intersection_faces, point_belongs
@@ -43,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ClosedPolygon2D: point_belongs.
 - Fullarc: get_reverse.
 - Arc2D: point_belongs
+- ArcEllipse2D: point_at_abscissa
 
 ### Refactor
 - ClosedShell3D: point_belongs, get_non_intersecting_faces
@@ -57,6 +59,7 @@ All adequations have been done for all tests and existing scripts.
 
 - bspline_compiled: refactor binomial_coefficient for performance.
 - Improve step translator.
+- Delete inplace methods: rotation, translation and frame_mapping
 - OpenShell3D: faces_graph.